// src/types/medicao.ts
export type Medicao = {
  date_demanda_ponta: string;
  horas_ponta: number;
  consumo_ponta: number;
  demanda_ponta: number;
  date_demanda_fora_ponta: string;
  horas_fora_ponta: number;
  consumo_fora_ponta: number;
  demanda_fora_ponta: number;
  date_demanda_capacitivo: string;
  horas_capacitivo: number;
  consumo_capacitivo: number;
  demanda_capacitivo: number;
  dia: string;
  id: number;
};